+++ contracts/test/TestPool.sol
@@ -5,7 +5,11 @@ pragma solidity ^0.6.8;
 import "../Pool.sol";
 
 contract TestPool is Pool {
-    function _getNumShares(uint amount, uint multiplier, uint price) public pure returns (uint) {
-      return getNumShares(amount, multiplier, price);
+    function _getNumShares(
+        uint256 amount,
+        uint256 multiplier,
+        uint256 price
+    ) public pure returns (uint256) {
+        return getNumShares(amount, multiplier, price);
     }
 }
